Best solution for comments in revise
This solution provides the possibility to navigate in segment
comments as you would in 'translate' path.

In order to do that, the confirm modal window has been suppressed
and should be replaced by a soft alert.

In this commit the '-comment' hash suffix is introduced, to allow
the history balloon to link properly to the segment with visible
comments. This will allow email links to work point to the comment
directly.

TODO: there's a bug when you click back to a translated segment
from a non translated one.