Adds REJECTED button to revision UI
Prior to this commit, the segment was automatically rejected when an
issue was submitted.

Now, submitting an issue no longer marks the segment as rejected.
Instead a reject button is added. In fact, segment can be approved
after an issue is submitted ( in case of neutral issues with 0 score).

A change to UI init was necessary to make shortcuts pluggable by review_improved
estension.