fix change of reviewed_words_count
This change fixes a bogus check on version number.

Introduce a check to consider unmodified ICEs only the locked ICEs.

Unlocked ICEs are ICE which are modifiable by the translator.
If the translator clicks on Translate for those, even if not changed,
those are considered to be included into revised words.