fix assignment of penalty_points on de-revision
This change removes a bogus query which is no longer necessary after
all. The problem was that the query which was responsibile to lookup for
penalty points to remove from chunk review, was no longer returning the
desired records because the issue is previously deleted.

Moreover, there's no longer need to query for penalty points to remove,
because we already run a query to find issues to delete. The sum of
penalty points to delete is now collected reducing the issue deletion
list array.